task-manager-backend/requirements.txt
python-dotenv==1.0.1
pymongo==4.6.2
dnspython==2.6.1
pyjwt==2.8.0
pydantic==2.6.4
bcrypt==4.1.2